Track dom changes and reapply our UI
This approach might need to be altered a bit. Right now it seems that we
are apply our changes to EACH element that gets added. If we need to
stick in a fetch request each time to get info for the added song then
we could be stacking like 80 fetch requests with one page reload... I
could see us avoiding this in two ways:
1. Cache the entirety of a users song / tag info on the front end (maybe
too costly, but maybe viable because it shouldn't actually be that much)
2. Try to bundle up all the rows of changes and do a single fetch for
the entire update (more ideal)